Run transport test in isolation
This test does some dark and dangerous mocking of static methods and
depends on the static initializer of some netty classes being run with a
system property set, so it really needs an isolated JVM to avoid
clashing with other tests.
