+++ kroxylicious-runtime/pom.xml
@@ -239,6 +239,30 @@
                         @{jacoco.argline} -javaagent:${org.mockito:mockito-core:jar} -Xshare:off
                     </argLine>
                 </configuration>
+                <executions>
+                    <execution>
+                        <id>run-isolated-test</id>
+                        <phase>test</phase>
+                        <goals>
+                            <goal>test</goal>
+                        </goals>
+                        <configuration>
+                            <forkCount>1</forkCount>
+                            <reuseForks>false</reuseForks>
+                            <groups>requiresIsolatedJvm</groups>
+                        </configuration>
+                    </execution>
+                    <execution>
+                        <id>default-test</id>
+                        <phase>test</phase>
+                        <goals>
+                            <goal>test</goal>
+                        </goals>
+                        <configuration>
+                            <excludedGroups>requiresIsolatedJvm</excludedGroups>
+                        </configuration>
+                    </execution>
+                </executions>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
@@ -281,9 +305,9 @@
                         </goals>
                         <phase>process-sources</phase>
                         <configuration>
-                        <directories>
-                            <directory>${project.build.directory}/generated-sources/krpc</directory>
-                        </directories>
+                            <directories>
+                                <directory>${project.build.directory}/generated-sources/krpc</directory>
+                            </directories>
                         </configuration>
                     </execution>
                 </executions>
@@ -299,7 +323,8 @@
                         </goals>
                         <phase>process-sources</phase>
                         <configuration>
-                            <messageSpecDirectory>${project.build.directory}/message-specs/common/message</messageSpecDirectory>
+                            <messageSpecDirectory>${project.build.directory}/message-specs/common/message
+                            </messageSpecDirectory>
                             <messageSpecFilter>*{Request,Response}.json</messageSpecFilter>
                             <templateDirectory>${project.basedir}/src/main/templates</templateDirectory>
                             <templateNames>BodyDecoder.ftl</templateNames>
@@ -316,7 +341,8 @@
                         </goals>
                         <phase>generate-sources</phase>
                         <configuration>
-                            <messageSpecDirectory>${project.build.directory}/message-specs/common/message</messageSpecDirectory>
+                            <messageSpecDirectory>${project.build.directory}/message-specs/common/message
+                            </messageSpecDirectory>
                             <messageSpecFilter>*{Request,Response}.json</messageSpecFilter>
                             <templateDirectory>${project.basedir}/src/main/templates</templateDirectory>
                             <templateNames>FilterInvoker.ftl</templateNames>
@@ -333,7 +359,8 @@
                         </goals>
                         <phase>process-sources</phase>
                         <configuration>
-                            <messageSpecDirectory>${project.build.directory}/message-specs/common/message</messageSpecDirectory>
+                            <messageSpecDirectory>${project.build.directory}/message-specs/common/message
+                            </messageSpecDirectory>
                             <messageSpecFilter>*{Request,Response}.json</messageSpecFilter>
                             <templateDirectory>${project.basedir}/src/main/templates</templateDirectory>
                             <templateNames>SpecificFilterArrayInvoker.ftl</templateNames>
